address review (codex): extract SSE_HEADERS to dependency-free sseHeaders.js
Importing SSE_HEADERS from sseDownload.js (which pulls in the HuggingFace
download stack via hfCache/hfDownload) dragged that heavy stack into every
sseUtils consumer at module load, breaking pipeline runner tests that mock
fileUtils (they failed at import before any test ran). Move SSE_HEADERS to a
new dependency-free server/lib/sseHeaders.js that both sseDownload.js and
sseUtils.js import; barrel + README updated.
