+++ server/lib/index.js
@@ -89,6 +89,7 @@ export * from './hfErrors.js';
 export * from './hfCache.js';
 export * from './hfDownload.js';
 export * from './sseDownload.js';
+export * from './sseHeaders.js';
 export * from './mediaModels.js';
 export * from './promptPartials.js';
 export * from './promptRunner.js';
