+++ server/lib/sseDownload.js
@@ -8,18 +8,9 @@
 
 import { inspectModelCache } from './hfCache.js';
 import { downloadHfRepo } from './hfDownload.js';
+import { SSE_HEADERS } from './sseHeaders.js';
 
 const inFlight = new Map(); // repo -> { promise, kill }
-
-export const SSE_HEADERS = {
-  'Content-Type': 'text/event-stream',
-  'Cache-Control': 'no-cache',
-  Connection: 'keep-alive',
-  // Disable proxy (nginx) response buffering so frames reach the client as they
-  // are written rather than being held until the buffer fills — without this an
-  // SSE stream behind a reverse proxy appears to hang.
-  'X-Accel-Buffering': 'no',
-};
 
 /**
  * Open an SSE response and return write-safe helpers. Canonical replacement
